requirements.txt
torch==2.0.1
tqdm==4.65.0
wandb==0.15.8
numpy==1.25.2
pillow==10.0.0
datasets==2.14.3
pywavelets==1.4.1
matplotlib==3.7.2
torchvision==0.15.2
torchmetrics==1.0.2
pytorch-wavelets==1.3.0
opencv-python==4.8.0.74
